requirements.txt
requests
openai
python-dotenv
alpaca_trade_api
python-dotenv
